@@ -2,13 +2,11 @@
 import 'core-js/stable'
 import 'regenerator-runtime/runtime'
 
-import Vue from 'vue'
-import vuetify from './plugins/vuetify'
+import { createApp, h } from 'vue'
+// import vuetify from './plugins/vuetify'
 import App from './app'
 
-Vue.config.productionTip = false
-
-new Vue({ // eslint-disable-line no-new
-  vuetify,
-  render: h => h(App)
-}).$mount('#app')
+createApp({
+  // vuetify,
+  render: () => h(App)
+}).mount('#app')
